Synesthesia3D: *Documentation update *Render targets are now stackable (make sure to disable them in the right order) *Fixed a bug where all render targets would have the automatic mip generation flag set
GITechDemo:
*Fixed resolution switching bug where not all resolutions would appear in the HUD on certain setups
*Build timestamp (visible in title of application window) is now forced to update as a pre-build event

General:
*Minor/latent bug fixes

+++ GITechDemo/Code/AppMain/GITechDemo/RenderScheme/SkyPass.cpp
@@ -69,7 +69,7 @@ void SkyPass::CreateSkyBoxVB()
 
 	const unsigned int ibIdx = ResourceMgr->CreateIndexBuffer(36);
 	IndexBuffer* ib = ResourceMgr->GetIndexBuffer(ibIdx);
-	const unsigned int sbIndices[] = {
+	const unsigned short sbIndices[] = {
 		0, 1, 2, 2, 1, 3,	// front
 		5, 4, 7, 7, 4, 6,	// back
 		4, 0, 6, 6, 0, 2,	// left
